backend/requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
djangorestframework-simplejwt>=5.3,<6.0
django-ratelimit>=4.1,<5.0
reportlab>=4.2,<5.0
python-dotenv>=1.0,<2.0
hypothesis>=6.100,<7.0
PyPDF2>=3.0,<4.0
gunicorn>=22.0,<23.0
qr_code
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1.0
whitenoise>=6.6.0
resend>=0.8.0
